+++ src/server/services/toolExecution/serverRuntimes/knowledgeBase.ts
@@ -53,8 +53,21 @@ export const knowledgeBaseRuntime: ServerRuntimeRegistration = {
         },
       },
       {
-        addFilesToKnowledgeBase: (knowledgeBaseId, ids) =>
-          knowledgeBaseModel.addFilesToKnowledgeBase(knowledgeBaseId, ids),
+        addFilesToKnowledgeBase: async (knowledgeBaseId, ids) => {
+          try {
+            return await knowledgeBaseModel.addFilesToKnowledgeBase(knowledgeBaseId, ids);
+          } catch (e: any) {
+            // PG unique-constraint violation on (knowledge_base_id, file_id).
+            // Re-throw with a friendly message so the ExecutionRuntime's
+            // generic catch surfaces it directly.
+            if (e?.cause?.code === '23505' || e?.code === '23505') {
+              throw new Error('One or more files are already in this knowledge base.', {
+                cause: e,
+              });
+            }
+            throw e;
+          }
+        },
         createKnowledgeBase: async ({ description, name }) => {
           const data = await knowledgeBaseModel.create({ description, name });
           return data?.id ?? '';
